requirements.txt
streamlit==1.53.1
psutil
requests
fastapi
uvicorn
python-dotenv